feat: add missing props & export all
